requirements.txt
streamlit
pandas
numpy
pdfplumber
openpyxl
google-generativeai>=0.5.0
python-dotenv
plotly
